Create Vsd (Vision Session Directory) 'nix/'nux library stub
Splitting the session directory parser across two foundation libraries
never made any sense.  Beginning refactoring that gives it it's own
library that will depend on the appropriate foundation libraries (V,
Vca, and Vsa) and be a dependent of any higher level library or tool
that needs access to the session directory (most of the higher level
tools and libraries).